# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cosco API Documentation",
        "description": "API documentation for Cosco application",
        "contact": {
            "email": "admin@cosco.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/wishlists": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get all wishlists for the authenticated user",
                "operationId": "978401ca2ca07b592255725cd25a443d",
                "responses": {
                    "200": {
                        "description": "List of wishlists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WishlistResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Create a new wishlist for the authenticated user",
                "operationId": "4bdf63de7775026077073b1069c45f64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreWishlistRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wishlist created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WishlistResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wishlists/{id}": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get a single wishlist for the authenticated user (with items)",
                "operationId": "b3cf1f24ca007a63db3f93f3e5e824d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wishlist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WishlistResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Update a wishlist for the authenticated user",
                "operationId": "d716f58b19d41444916708b6e9eac2c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wishlist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateWishlistRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wishlist updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WishlistResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreWishlistRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "لیست علاقه‌مندی‌ها"
                    }
                },
                "type": "object"
            },
...
            "UpdateWishlistRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "لیست علاقه‌مندی‌ها"
                    }
                },
                "type": "object"
            },
...
            "WishlistItemResource": {
                "title": "Wishlist Item Resource",
                "description": "Wishlist item resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "product_slug": {
                        "type": "string",
                        "example": "product-slug"
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Product Name"
                    },
                    "product_image_url": {
                        "type": "string",
                        "example": "https://example.com/product.jpg"
                    },
                    "product_price": {
                        "type": "string",
                        "example": "100,000 تومان"
                    }
                },
                "type": "object"
            },
            "WishlistResource": {
                "title": "Wishlist Resource",
                "description": "Wishlist resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "لیست علاقه‌مندی‌ها"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/WishlistItemResource"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}